[PUBDEV-3616] Move Assembly.java to water.codegen
