+++ h2o-core/src/main/java/water/codegen/Assembly.java
@@ -1,4 +1,4 @@
-package water.rapids;
+package water.codegen;
 
 import water.H2O;
 import water.Key;
